find_a_program: Only search for prefixed paths in undisambiguated dirs
This means, we might search for:

- path/$machine/$version/prog
- path/$machine/prog
- path/$machine-prog

But not

- path/$machine/$version/$machine-prog

because disambiguating $machine twice is unnecessary.

This does mean we less liberal in what we accept than LLVM, but that's
OK.

(The down side of always following Postel's law is that everyone
converges on accepting all sorts of garbage, which makes debugging
end-to-end hard when mistakes are not caught early.)

In plainer language: while I am citing LLVM as precedent, I am
conservatively only adopting the parts of LLVM's behavior that I
actually need, and which I think is the most justifiable. It is easier
to further liberalize the search behavior later (always look for
prefixed and unprefixed, lets say), than tighten it up (creating a
breaking change).

	* gcc.cc (program_at_path): Use the boolean parameter added in the
	previous commit to implement the above logic.

Signed-off-by: John Ericson <git@JohnEricson.me>