Allow explicitly specifying the thread model for runtime libs
Previously, they always scraped the thread mode from `$CC -v', now, that
is the default but one may pass `gcc_cv_target_thread_file=MODEL` to be
explicit instead.

The use-case is for making multilib builds where the library sets do not
all share the same thread model easier. (See also `THREAD_MODEL_SPEC`
for this use-case.)

Arguably we could try to define on `THREAD_MODEL_SPEC` on more platforms
(besides e.g. AIX) but the ramifications of this are a bit unclear. Even
if that does become the best way to do this, better than setting
`gcc_cv_target_thread_file` directly, the macro this introduced to
deduplicate the `$CC -v' scraping is probably still a worth having.

Suggested-by: Arsen Arsenović <arsen@aarsen.me>