Use regular libdir/includedir in libgcc Makefiles
My long-term goal is to get all "libsubdir" and similar logic out of the
build systems for the runtime libraries themselves, and instead move it
to the "meta build system". The idea is the runtime library builds
shouldn't just be given the locations they are to install stuff in, and
don't need to know or care how those locations were calculated --- it is
just better separation of concerns.

Another motivation is for the new Nixpkgs packaging of GCC I am working
on. Nix already installs everything to separate prefixes, so the
libsubdir logic that GCC is doing is redundant --- we already have no
conflicts by construction and are passing many -B flags to make things
go anyways. This separation of concerns makes it easier to build runtime
libraries (which are already doing separately in the new packaging) with
"normal" libdir/subdir.

This current patch doesn't go as far as removing all such subdir logic
from the libgcc build --- rather, as an intermediate goal, it moves it
out of the Makefiles and into the configure script. That is to say, in
the makefiles there is no longer any more libsubdir, and everything just
uses libdir/includedir, while in the configure script, the original
libdir is assigned to a temporary (non-substituted) org_libdir variable,
and then libdir and includedir are redefined to be the proper subdirs of
org_libdir.

For what it's worth, this intermediate goal does match how slibdir,
toolexecdir, and toolexeclibdir are also calculated in the configure
script from flags and "standard" autoconf directory variables. So even
though the main goal is not fully realized, I do think the patch is
already lending the build system a bit more consistency as an added
benefit.

libgcc/ChangeLog:

	* Makefile.in: Use regular libdir/includedir, no more libsubdir
	etc.
	* config/t-slibgcc-vms: Use regular libdir, no more libsubdir.
	* configure: Regenerate.
	* configure.ac: Stash away the passed libdir as orig_libdir, and
	define libdir and includedir to matches what Makefile.in use to
	do.

Signed-off-by: John Ericson <git@JohnEricson.me>