Use regular libdir/includedir in libssp Makefiles
The details are exactly the same as
https://gcc.gnu.org/pipermail/gcc-patches/2025-July/689865.html, except
this is for libssp not libgcc.

It probably makes sense to just pick one of these to review first, and
then based on any feedback I can rework the other one. But for
posterity, and to illustrate the pattern, I am submitting this one too
before the other is submitted.

libssp/ChangeLog:

	* Makefile.am: Use regular libdir/includedir, no more libsubdir
        etc.
	* Makefile.in: Regenerate.
	* configure: Regenerate.
	* configure.ac: Stash away the passed libdir as orig_libdir, and
        define libdir and includedir to match what Makefile.in used to
        do.

Signed-off-by: John Ericson <git@JohnEricson.me>