driver: Rework for_each_path using C++
The old C-style was cumbersome make making one responsible for manually
create and pass a closure (separate function and *_info class for
closed-over variables).

I would have liked to redo this with C++ lambdas, so we could:

- derive environment types implicitly
- have fewer stray static functions
- make the return-type polymorphism type safe.

But I was advised on IRC that introducing template specializations in
the driver would a open compile time performance can of worms. It was
instead suggested I do it the old OOP way: this patch.

The old OOP way I don't really like, as it doesn't really reduce
boilerplate, but at least the environment accessing is well-typed now
(no `data` to `info` casts), and the `info->` boilerplate is also gone.
But I am happy to do as suggested. Also I have an idea for a follow-up
patch (after this one) that does allow the use of lambdas without
templating the original function, and so should be a "best of both
worlds" approach.

Thanks to Jason Merill for the idea of using an out-of-line definition
to avoid a large reindentating. That does make this much nicer.

gcc/ChangeLog:

	* gcc.cc (for_each_path): Made method on class below.
	(class find_within_paths): Class below has virtual method for
	callback, instead of function and and data arguments.
	(struct add_to_obstack_info): Renamed without trailing _info.
	(struct add_to_obstack): Now inherits from find_within_paths.
	(add_to_obstack): Became the callback on the above class which
	inherited its name.
	(build_search_list): Updated to use add_to_obstack.
	(struct file_at_path_info): Renamed without trailing _info.
	(struct file_at_path): Now inherits from find_within_paths.
	(file_at_path): Became the callback on the above class which
	inherited its name.
	(find_a_file): Updated to use new file_at_path.
	(struct spec_path_info): Renamed without trailing _info.
	(struct spec_path): Now inherits from find_within_paths.
	(spec_path): Became the callback on the above class which
	inherited its name.
	(do_spec_1): Updated to use new spec_path.

Signed-off-by: John Ericson <git@JohnEricson.me>