driver: `add_to_obstack` and `find_a_file` redo with lambdas
A follow-up from "driver: Rework for_each_path using C++"

These types are, for all intents and purposes, single-use closure
environment types. It is much more ergonomic to juse use lambdas for
this.

On IRC there was concern about static dispatch and compile times with
lambdas, but I have kept the class hierarchy and virtual methods from
the previous commit that keeps `find_within_paths::for_each_path`
untemplated, and I believe that add addresses the issue.

Note that `struct spec_path` was *not* converted because it is used
multiple times. We could stil convert to a lambda, but we would want to
put the for_each_path call with that lambda inside a separate function
anyways, to support the multiple callers. Unlike the other two
refactors, it is not clear that this one would make anything shorter.

gcc/ChangeLog:

	* gcc.cc (struct add_to_obstack): Gone, lambda capture list
	replaces it.
	(struct file_at_path): Gone, lambda capture list replaces it.
	(add_to_obstack::callback): Moved to lambda in
	build_search_list.
	(file_at_path::callback): Moved to lambda in fine_a_file.
	(for_each_path): New tiny templated wrapper for lambda
	(build_search_list): Use for_each_path with lambda.
	(find_a_file): Use for_each_path with lambda.

Signed-off-by: John Ericson <git@JohnEricson.me>